Move Command to hikari.commands
