+++ tests/hikari/interactions/test_command_interactions.py
@@ -35,90 +35,6 @@ def mock_app():
     return mock.Mock(traits.CacheAware, rest=mock.AsyncMock())
 
 
-class TestCommand:
-    @pytest.fixture()
-    def mock_command(self, mock_app):
-        return command_interactions.Command(
-            app=mock_app,
-            id=snowflakes.Snowflake(34123123),
-            application_id=snowflakes.Snowflake(65234123),
-            name="Name",
-            description="very descript",
-            options=[],
-            guild_id=snowflakes.Snowflake(31231235),
-        )
-
-    @pytest.mark.asyncio()
-    async def test_fetch_self(self, mock_command, mock_app):
-        result = await mock_command.fetch_self()
-
-        assert result is mock_app.rest.fetch_application_command.return_value
-        mock_app.rest.fetch_application_command.assert_awaited_once_with(65234123, 34123123, 31231235)
-
-    @pytest.mark.asyncio()
-    async def test_fetch_self_when_guild_id_is_none(self, mock_command, mock_app):
-        mock_command.guild_id = None
-
-        result = await mock_command.fetch_self()
-
-        assert result is mock_app.rest.fetch_application_command.return_value
-        mock_app.rest.fetch_application_command.assert_awaited_once_with(65234123, 34123123, undefined.UNDEFINED)
-
-    @pytest.mark.asyncio()
-    async def test_edit_without_optional_args(self, mock_command, mock_app):
-        result = await mock_command.edit()
-
-        assert result is mock_app.rest.edit_application_command.return_value
-        mock_app.rest.edit_application_command.assert_awaited_once_with(
-            65234123,
-            34123123,
-            31231235,
-            name=undefined.UNDEFINED,
-            description=undefined.UNDEFINED,
-            options=undefined.UNDEFINED,
-        )
-
-    @pytest.mark.asyncio()
-    async def test_edit_with_optional_args(self, mock_command, mock_app):
-        mock_option = object()
-        result = await mock_command.edit(name="new name", description="very descrypt", options=[mock_option])
-
-        assert result is mock_app.rest.edit_application_command.return_value
-        mock_app.rest.edit_application_command.assert_awaited_once_with(
-            65234123, 34123123, 31231235, name="new name", description="very descrypt", options=[mock_option]
-        )
-
-    @pytest.mark.asyncio()
-    async def test_edit_when_guild_id_is_none(self, mock_command, mock_app):
-        mock_command.guild_id = None
-
-        result = await mock_command.edit()
-
-        assert result is mock_app.rest.edit_application_command.return_value
-        mock_app.rest.edit_application_command.assert_awaited_once_with(
-            65234123,
-            34123123,
-            undefined.UNDEFINED,
-            name=undefined.UNDEFINED,
-            description=undefined.UNDEFINED,
-            options=undefined.UNDEFINED,
-        )
-
-    @pytest.mark.asyncio()
-    async def test_delete(self, mock_command, mock_app):
-        await mock_command.delete()
-
-        mock_app.rest.delete_application_command.assert_awaited_once_with(65234123, 34123123, 31231235)
-
-    @pytest.mark.asyncio()
-    async def test_delete_when_guild_id_is_none(self, mock_command, mock_app):
-        mock_command.guild_id = None
-
-        await mock_command.delete()
-
-        mock_app.rest.delete_application_command.assert_awaited_once_with(65234123, 34123123, undefined.UNDEFINED)
-
-
 class TestCommandInteraction:
     @pytest.fixture()
     def mock_command_interaction(self, mock_app):
